move go artifacts to api-client-staging
Many languages put the generated artifacts in the google-cloud-{lang}
directory under REPOROOT.
One unusual thing about Go is that the toolchain assumes all Go code to
live inside a directory named by env var GOPATH.
This path is different on each machine,
so we don't have a good default place to put the artifacts.

Knowing this, we might as well make it easier to put things into
staging repo.