feat: Add Resource to the vp1beta1 NotificationMessage, which contains resource-related fields associated to the Finding notification.
feat!: Removed 'having' as a field for the Group and List APIs for both Assets and Findings.

docs: Adjusted wording from Cloud SCC to Security Command Center, and updated filtering docs for List APIs.

BREAKING CHANGE: Removed 'having' as a field in the List and Group Finding and Asset requests.

PiperOrigin-RevId: 315725809