fix: Specify namespaces for C# and PHP
Note: this isn't marked as a breaking change in the conventional commit as it is *only* breaking for C# and PHP, and Google Cloud Client Libraries for the Access Approval API have not been released in these languages yet. Anyone who previously generated libraries for these languages will see a breaking change when they regenerate though.
PiperOrigin-RevId: 345470945