chore: Make microgenerators default for all languages
In particular this applies to Java and Python microgenerators.
With this change `java_gapic_*` and `py_gapic_*` rules mean microgenerator versions. The monolith versions of the rules are named as `java_gapic_*_legacy` and `py_gapic_*_legacy` respectively.

Also change Ads BAZEL template to reflect the change in default rule names.

Also change `DO NOT OVERRIDE` comment warnings meaning (to reflect that now by default microgenerator rules are generated and the generated version must be accepted only if the API is being migrated to the microgenerator).

PiperOrigin-RevId: 346227314