fix: specify valid go_package option
The explicitly specified package name is invalid since it identical
to one of the keywords in Go (e.g., "package" or "type").
Remove explicitly specifying the name, which will cause protoc-gen-go
to derive the package name from the last path segment.
If the name is not valid, it will automatically sanitize it
to produce a valid package identifier (e.g., "_package" or "_type").

Committer: @dsnet
PiperOrigin-RevId: 370589666