fix: Remove invalid resource annotations
Note: normally removing an annotation is a breaking change. However, the annotation here is invalid as it doesn't refer to a string field. Any generator which actually tried to use it in a meaningful way would either generate invalid code or fail to generate at all. Therefore removing it can't break anything.
PiperOrigin-RevId: 404455880