feat!: modified descriptions of validation rules on create entity endpoints * Commas disallowed * Backslash allowed
BREAKING CHANGE: Because of a bug commas have been accepted as part of individual IDs. Fixing this bug is a breaking change but we don't expect users to be impacted as we have detected no individual IDs being generated with commas. Commas are still OK to use as delimiters for lists of multiple IDs.

PiperOrigin-RevId: 431967940