feat: add new toll passes feat!: add optional tags to proto3 index based fields to enable checking for field presence
BREAKING CHANGE: Adds field presence to fields with semantic meanings for their default values. In this case index fields with a value of 0 have specific meaning and their presence should be noted. This may break workflows where client libraries generated from this change handle data from previous client libraries (where presence is not noted). There are no users of the client libraries currently, so while this is a potential breaking change - there is nothing to break.

PiperOrigin-RevId: 470006887