feat: Add field for enabled Beta K8s APIs in proto
---
feat: add etags for cluster and node pool update operations

Clients may now optionally specify the previously read etag for a cluster or node pool during update requests. The etag is regenerated every time the object is updated. If an etag is provided, it will be checked against the current etag, and the request will be rejected on mismatch. This is particularly useful for allowing two clients to concurrently update a repeated field without overwriting each other.

PiperOrigin-RevId: 500787735