feat!: remove unused protocol buffers
This PR cleans up an experiment where we attempted to leverage protocol buffer definitions to simplify processing of the bigquery service API responses related to ML models.  There are no client libraries that consume this information, and the definitions cannot be used to successfully build any form of client library on their own.

PiperOrigin-RevId: 525854672