feat: add `model` and `language_codes` fields in `RecognitionConfig` message + enable default `_` recognizer
Enables specifying `model` and `language_codes` in requests without having to specify them in the Recognizer (they can still be specified in the Recognizer in the `default_recognition_config` field).

Also enables using the recognizer ID `_` to perform recognition without explicitly creating a Recognizer resource.

The top-level `model` and `language_codes` fields are deprecated in favor of the new fields added in the `RecognitionConfig` message. The old fields continue to work.

PiperOrigin-RevId: 545698919